small compilation fix

+++ sql/backends/monet5/wlr.c
@@ -908,11 +908,11 @@ WLRappend(Client cntxt, MalBlkPtr mb, MalStkPtr stk, InstrPtr pci)
 	}
 
 	if (cname[0] != '%' && (c = mvc_bind_column(m, t, cname)) != NULL) {
-		store_funcs.append_col(m->session->tr, c, pos, ins, TYPE_bat);
+		store_funcs.append_col(m->session->tr, c, (size_t)pos, ins, TYPE_bat);
 	} else if (cname[0] == '%') {
 		sql_idx *i = mvc_bind_idx(m, s, cname + 1);
 		if (i)
-			store_funcs.append_idx(m->session->tr, i, pos, ins, tpe);
+			store_funcs.append_idx(m->session->tr, i, (size_t)pos, ins, tpe);
 	}
 cleanup:
 	BBPunfix(((BAT *) ins)->batCacheid);
